pyproject.toml
[project]
name = "asistente-api"
version = "0.1.0"
description = "Aplicacion RAG y API para Asistente Legal de transito bolivia"
authors = [
    {name = "Cesar Caballero",email = "udemicscesar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.22,<0.4.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "openai (>=1.70.0,<2.0.0)",
    "poetry-plugin-shell (>=1.0.1,<2.0.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
